Add WGPU procs and WGPUGetProcAddress.
Having a GetProcAddress mechanism will be important to support
extensions that expose new functions. Procs are defined for all
entry-points because from experience in EGL_KHR_get_all_proc_addresses
developers want to sometimes be able to query all functions through
GetProcAddress.

WGPUGetProcAddress takes an optional WGPUDevice argument so that, like
in Vulkan, the procs can be specialized for a given device.