Add wgpuDeviceDestroy
This is the equivalent of wgpuDeviceDestroy that allows the device to
free all resource allocations and abort asynchronous operations
immediately. Basically it forces a fraceful immediate loss.

However like the JS API, it doesn't actually destroy the CPU-side
objects that can still be refcounted / freed, and must all be freed
otherwise memory will leak.