Add limits
Limits are represented by an enum WGPULimitName.

Limits can be queried using wgpuAdapterGetLimitU32 or with
wgpuAdapterGetLimitU64, depending on the type of the limit.
These functions return false if the implementation does not
recognize the limit name, or if the limit type does not match.

Limits can be requested by passing a list of WGPULimitEntry in
the WGPUDeviceDescriptor. Values are all uint64_t, but the
implementation validates that they are within the range of the
expected limit type.