Add base WGPULimits structs (#104)
This struct may have additional chained structs to request additional
limits that are added to the API or are present in optional features.

An adapter's supported limits may be queried using wgpuAdapterGetLimits.
This function populates the provided limits and any chained limit
structs. It is invalid to pass chained limit structs here for features
that are not supported by the adapter.

requiredLimits may be passed in the WGPUDeviceDescriptor to request
specific non-default limits. Since this is a struct, there is no way to
optionally include members. #defines WGPU_LIMIT_{type}_UNDEFINED may be
used to omit a limit from the request and get the default instead.
It is invalid to pass chained limit structs here for features that are
not supported by the adapter.