Add a way to enumerate available features
Adds wgpuAdapterEnumerateFeatures and wgpuDeviceEnumerateFeatures
to enumerate features available on the adapter or enabled on the
device. Returns `count` where `count` is the number of features.
Calling with non-null will write out `count` values sequentially
starting at the pointer address provided.

Also adds wgpuDeviceHasFeature which was previously added to the
adapter, but missed on the device.

For #129. This makes it easier to implement the setlike interface
in the JS API which is iterable, without checking HasFeature on
every known feature.