Document wgpuGetProcAddress, copydoc for all Proc definitions
Instead of duplicating docs, have each proc type just point to the
appropriate definition and quote its docs.

Doxygen doesn't actually generate docs for the procs, I guess because
they're not in a \defgroup, but it works if I do add a \defgroup.
Regardless, this should hopefully work in IDEs supporting doc-comments.

Fixes 315