Rename WGPUProgrammableStageDescriptor -> WGPUComputeStage
In JS, `WGPUProgrammableStage` (no `Descriptor`, it was renamed) is the
base dictionary for vertex/fragment/compute. But in C we don't use it
for that, we only use it for compute because compute doesn't add any
other members. So we should name it accordingly. Since the JS type is
`WGPUProgrammableStage`, I've called it `WGPUComputeStage`.

Both `WGPUComputePipelineDescriptor` and `WGPUComputeStage` are still
both extensible, which makes reasonable sense and matches
`WGPURenderPipelineDescriptor`/`WGPUVertexState`/`WGPUFragmentState`.

Fixes 409