Rename WGPUInstanceFeatureName_TimedWaitAnyEnable -> TimedWaitAny (#549)
`WGPUInstanceFeature_TimedWaitAnyEnable` (formerly
`WGPUInstanceCapabilities.timedWaitAnyEnable`) should be called simply
`TimedWaitAny` now that it's a feature name.

Aside, it's arguably redundant with `timedWaitAnyMaxCount`. It would be
simpler to use if we didn't have a feature, and just set
`timedWaitAnyMaxCount` to 0 or not. The reason this wasn't the case
originally is because it has a minimum value of 64, so if we did this
then:
- 0 -> disabled with a max of 0
- 1 -> enabled with a max of 64

which is slightly odd. I'm inclined not to bother to change this.