[minor breaking] Add new feature names from upstream JS API (#551)
These were added in the upstream WebGPU API and should be reflected
here.
- https://github.com/gpuweb/gpuweb/pull/5147
Ordered at the top for consistency even though it changes the numbering
of everything else.
Also fixed the ordering of TimestampQuery to match since the numbering
is changing anyway. (Even though the ordering doesn't really matter
because it's nonobservable upstream and there isn't really an ordering.)
- https://github.com/gpuweb/gpuweb/pull/5160
- https://github.com/gpuweb/gpuweb/pull/5226

The renumbering is technically a breaking change because it changes the
ABI (which shouldn't matter right now) and could affect code that
depends on comparing enum values or hard-coded integer versions of them.