Add stub ExternalTexture and support for them in BindGroups (#570)
PTAL! It doesn't add creation of the external texture as that's still
super implementation specific, but this will unblock folks that want to
use ExternalTexture in WASM (they can import a C object from a
precreated JS GPUExternalTexture).

The webgpu.h diff is slightly messy because the new chained struct force
a small reordering. The additions themselves should be pretty trivial
and they add a chained struct for the equivalent Rust enum option in
wgpu.

I chose `ExternalTextureBindingLayout` to match the
`Buffer/Sampler/...BindingLayout` convention that's already used for
`BindGroupLayoutEntry` members. `ExternalTextureBindingEntry` is just
following that. But I'm not to attached to the exact names.